refactored get defect translation to defect repository
